requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
discord-pretty-help==1.3.1
discord-py-slash-command==1.2.2
discord.py==1.7.3
dpymenus==2.1.5
dpytools==0.18.0b0
emoji==1.2.0
idna==2.10
motor==2.3.1
multidict==5.1.0
odmantic==0.3.5
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
PyNaCl==1.4.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3